chore: altera caminho do notebook
